tests/wpt/web-platform-tests/html/webappapis/dynamic-markup-insertion/document-write/module.html

<!doctype html>
<meta charset=utf-8>
<title>document.write in a module</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
async_test(t => {
  const iframe = document.createElement("iframe");
  iframe.src = "module-iframe.html";
  document.body.appendChild(iframe);
  iframe.onerror = t.unreached_func("Error loading iframe");
  iframe.onload = t.step_func_done(() => {
    assert_equals(iframe.contentDocument.body.textContent, "PASS\n");
  });
});
</script>
